Smart Chatbot_Gemma AI/requirements.txt
# conda create -p venv python==3.10
# pip install -r requirements.txt
# conda activate venv/
langchain
python-dotenv
langchain-openai
ipykernel
beautifulsoup4
langchain_community
faiss-cpu
streamlit